+++ src/application-container/index.js
@@ -1,9 +1,11 @@
+import ApplicationContainer from './ApplicationContainer';
 import useActiveMainPage from './useActiveMainPage';
 import ApplicationContainerContext, {
   useApplicationContainer,
   contextShape as applicationContainerContextShape,
 } from './ApplicationContainerContext';
 
+export default ApplicationContainer;
 export {
   useActiveMainPage,
   ApplicationContainerContext,