+++ src/application-container/index.js
@@ -1,11 +1,20 @@
+import ApplicationContainer from './ApplicationContainer';
 import useActiveMainPage from './useActiveMainPage';
+import ApplicationContainerContext, {
+  useApplicationContainer,
+  contextShape as applicationContainerContextShape,
+} from './ApplicationContainerContext';
 import ApplicationConceptContext, {
   useApplicationConcept,
   contextShape as applicationConceptContextShape,
 } from './ApplicationConceptContext';
 
+export default ApplicationContainer;
 export {
   useActiveMainPage,
+  ApplicationContainerContext,
+  useApplicationContainer,
+  applicationContainerContextShape,
   ApplicationConceptContext,
   useApplicationConcept,
   applicationConceptContextShape,